Added matrices file via upload
Experimenting with 2 dimensional arrays and for loops in this one, getting better at actually returning 0 in the main function, telling the program to stop running. Most of the programming I've done has been in higher level languages, so some things in C are hard to get a grasp on. Also, going to be adding exe files for these now.